Create packetbeat x-pack magefile
